Flatbuffer based serialization of Tensor
