+++ ttnn/cpp/ttnn/tensor/flatbuffer/mesh_coord_flatbuffer.cpp
@@ -1,0 +1,35 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "mesh_coord_flatbuffer.hpp"
+
+#include <tt-metalium/mesh_coord.hpp>
+
+#include <vector>
+
+namespace ttnn {
+
+flatbuffers::Offset<flatbuffer::MeshCoordinate> to_flatbuffer(
+    const tt::tt_metal::distributed::MeshCoordinate& coord, flatbuffers::FlatBufferBuilder& builder) {
+    auto values_vector = builder.CreateVector(std::vector<uint32_t>(coord.coords().begin(), coord.coords().end()));
+    return flatbuffer::CreateMeshCoordinate(builder, values_vector);
+}
+
+tt::tt_metal::distributed::MeshCoordinate from_flatbuffer(const flatbuffer::MeshCoordinate* coord) {
+    return tt::tt_metal::distributed::MeshCoordinate(
+        std::vector<uint32_t>(coord->values()->begin(), coord->values()->end()));
+}
+
+flatbuffers::Offset<flatbuffer::MeshShape> to_flatbuffer(
+    const tt::tt_metal::distributed::MeshShape& shape, flatbuffers::FlatBufferBuilder& builder) {
+    auto dimensions_vector = builder.CreateVector(std::vector<uint32_t>(shape.view().begin(), shape.view().end()));
+    return flatbuffer::CreateMeshShape(builder, dimensions_vector);
+}
+
+tt::tt_metal::distributed::MeshShape from_flatbuffer(const flatbuffer::MeshShape* shape) {
+    return tt::tt_metal::distributed::MeshShape(
+        std::vector<uint32_t>(shape->dimensions()->begin(), shape->dimensions()->end()));
+}
+
+}  // namespace ttnn
