+++ ttnn/cpp/ttnn/tensor/flatbuffer/mesh_coordinate.fbs
@@ -1,0 +1,8 @@
+namespace ttnn.flatbuffer;
+
+table MeshCoordinate {
+    // The coordinate values in each dimension
+    values:[uint32];
+}
+
+root_type MeshCoordinate;
