+++ ttnn/cpp/ttnn/tensor/flatbuffer/tensor_spec_flatbuffer.hpp
@@ -1,0 +1,22 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "tensor_spec_generated.h"
+
+#include "ttnn/tensor/types.hpp"
+#include "ttnn/tensor/tensor_spec.hpp"
+
+namespace ttnn {
+
+flatbuffers::Offset<flatbuffer::MemoryConfig> to_flatbuffer(
+    const tt::tt_metal::MemoryConfig& config, flatbuffers::FlatBufferBuilder& builder);
+flatbuffers::Offset<flatbuffer::TensorSpec> to_flatbuffer(
+    const tt::tt_metal::TensorSpec& spec, flatbuffers::FlatBufferBuilder& builder);
+
+tt::tt_metal::MemoryConfig from_flatbuffer(const flatbuffer::MemoryConfig* config);
+tt::tt_metal::TensorSpec from_flatbuffer(const flatbuffer::TensorSpec* spec);
+
+}  // namespace ttnn
